+++ mqttwarn/services/apprise_multi.py
@@ -5,10 +5,13 @@ __copyright__ = 'Copyright 2021 Andreas Motl'
 __license__   = 'Eclipse Public License - v 1.0 (http://www.eclipse.org/legal/epl-v10.html)'
 
 # https://github.com/caronc/apprise#developers
-from urllib.parse import urlencode
 from collections import OrderedDict
 
 import apprise
+
+from mqttwarn.services.apprise_util import obtain_apprise_arguments, add_url_params, get_all_template_argument_names
+
+APPRISE_ALL_ARGUMENT_NAMES = get_all_template_argument_names()
 
 
 def plugin(srv, item):
@@ -39,6 +42,10 @@ def plugin(srv, item):
             # Collect URL parameters.
             params = OrderedDict()
 
+            # Obtain and apply all possible Ntfy parameters from data dictionary.
+            params.update(obtain_apprise_arguments(item, APPRISE_ALL_ARGUMENT_NAMES))
+
+            # Apply addressee information.
             if "recipients" in address:
                 to = ','.join(address["recipients"])
                 if to:
@@ -49,10 +56,8 @@ def plugin(srv, item):
             if "sender_name" in address:
                 params["name"] = address["sender_name"]
 
-            # Add notification services by server url.
-            uri = baseuri
-            if params:
-                uri += '?' + urlencode(params)
+            # Add parameters to Apprise notification URL.
+            uri = add_url_params(baseuri, params)
             srv.logging.info("Adding notification to: {}".format(uri))
             apobj.add(uri)